+++ ml_workflow.txt
@@ -31,13 +31,9 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
 
-      - name: Run pre model training tests
+      - name: Run model training unit tests
         run: |
-          python -m pytest tests/unit_test_train_model_methods.py
-
-      - name: Run post model training tests
-        run: |
-          python -m pytest tests/unit_test_train_model_saving.py
+          python -m pytest tests/unit_test_train_model.py
 
   CD:
     needs: CI
